pyproject.toml
[project]
name = "upb-rag-careers"
version = "0.1.0"
description = "RAG system for career exploration at UPB Medellín"
readme = "README.md"
requires-python = ">=3.12,<3.13"

dependencies = [
    "bs4>=0.0.2",
    "langchain==1.0.2",
    "langchain-classic==1.0.0",
    "langchain-community==0.4.0", # explicitly pinned to prebuilt wheel
    "langchain-openai>=0.3.14",
    "langchain-text-splitters>=0.3.4",
    "playwright>=1.55.0",
    "python-dotenv>=1.0.0",
    "faiss-cpu>=1.9.0",
    "rank-bm25>=0.2.2",  # For BM25 sparse retrieval
]
